feat: button object with image and loading bar selection
